#15
Page updated but similar to other tickets I have submitted today, I am
unable to implement the "formaction" piece of the form code at this
time. I have commented the file with the appropriate change to implement
once the controller is ready for it. (I also saw that someone else was
creating buttons differently than I did for my account page, so changing
it to match the other style is the update on that file.)