Aesthetics
I made some usability updates. I edited some of the titles of text entry
fields on the event creation page for user clarity, added some spacing
to the top of the account page to give it a uniform look with the rest
of the app, and removed the coloring from the promote and demote buttons
on the club roster as 3 colored buttons next to each other is against
best practice.