Frontend new colors! (#217)
* #209 hide the ID

* Applied ClubSuite color scheme to Bulma

Added local Bulma.css file and updated references.  Bulma.css is
compiled dynamically from the bulma.sass files using a sass compiler
(Koala).  Color scheme is controlled in the utilities/variables file in
the colors at the top.

* Aesthetics

I made some usability updates. I edited some of the titles of text entry
fields on the event creation page for user clarity, added some spacing
to the top of the account page to give it a uniform look with the rest
of the app, and removed the coloring from the promote and demote buttons
on the club roster as 3 colored buttons next to each other is against
best practice.

* add confirmation and comment out order